Basis setup!
Please note that from now on NALA cannot be run anymore without a proper project file fully set up in JCR6.
This does not yet exist, as no NALA projects are currently created (will be done soon, and the Bubble Builder will be able to assemble/compile them properly for NALA).

The first project planned is the revamp of Dyrt, but there is still a lot to be done there....