Direct X temporary version set up
For some odd reason this game no longer works in OpenGL.
This is a bug in MonoGame, not in my code. How do I know so sure? Because even base templates without any modifications at all refuse to work, and the error I get is the same. That's why.
In DirectX the engine DOES work, although not perfect. I will try to see if the game keeps compiling in OpenGL though, this to see if I can at least make Linux exports possible (no promises).

And this entire incident once again proves that I am in a desparate need of my very own engine with as little 3rd party crap as possible, as this incident almost killed tons of work....

1000 curses and foul words!